python/requirements.txt
# Healthcare Imaging MLOps Platform - Python Dependencies

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# SageMaker
sagemaker>=2.200.0
sagemaker-experiments>=0.1.45

# Machine Learning
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0
SimpleITK>=2.3.0
pydicom>=2.4.0
opencv-python>=4.8.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data Validation
marshmallow>=3.20.0
jsonschema>=4.20.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
moto>=4.2.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.0
tqdm>=4.66.0

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
